[PIDM-329] fix: integration tests
add channel utils page to reduce test suite complexity;
fix sonarqube static analysis;
